requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
firebase-admin==6.5.0
pydantic==2.7.1
python-multipart==0.0.9
apscheduler==3.10.4
httpx==0.27.0
